file restructure
errors
utilities
more utils
page layout
data management no include updates
particles no include updates
figures no include updates
sequencer no include updates
audio components no include updates
fix include paths
