+++ Source/AudioComponents/Cursor.cpp
