+++ Source/AudioComponents/Cursor.h
