+++ Source/AudioComponents/TracktionThumbnail.cpp
