+++ Source/AudioComponents/TracktionThumbnail.h
