+++ Source/Errors/ErrorManager.cpp
