+++ Source/Figures/FigureGenerator.cpp
