+++ Source/PageLayout/Figures.cpp
